+++ app/src/modules/settings/routes/translation-strings/translation-strings-drawer.vue
@@ -51,7 +51,7 @@ import { ref, computed, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { isEqual } from 'lodash';
 import { Field, DeepPartial } from '@directus/shared/types';
-import { useTranslationStrings, TranslationString } from '../../composables/use-translation-strings';
+import { useTranslationStrings, TranslationString } from '@/composables/use-translation-strings';
 
 interface Props {
 	modelValue: boolean;
